apps/voice-orchestrator-pipecat/requirements.txt
# Pipecat Voice AI Framework (without daily - we use Twilio)
pipecat-ai[azure,openai,silero]==0.0.54

# Twilio integration
twilio>=9.0.0
python-dotenv>=1.0.0

# FastAPI for webhooks
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP client for core-api
httpx>=0.26.0
aiohttp>=3.9.0

# Azure services
azure-cognitiveservices-speech>=1.35.0
openai>=1.12.0

# Utilities
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging
loguru>=0.7.2

